requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.4.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-CKEditor==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.2.1

python-dotenv~=1.0.1